backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-socketio==5.11.3
eventlet==0.36.0